+++ src/components/common/register.vue
@@ -1,5 +1,0 @@
-<template>
-  <button class="button green" disabled>
-    <strong>{{ $t('components.register.register_text') }}</strong>
-  </button>
-</template>