+++ src/components/navigation_bar.vue
@@ -1,10 +1,13 @@
 <template>
   <nav class="navbar is-fixed-top" role="navigation" aria-label="main navigation">
     <div class="navbar-brand noselect">
-      <!-- svg is required here!! -->
-      <!-- <router-link class="navbar-item" to="/">
-        <img :src="require('@/assets/Logo.png')" alt="localzon(e)">
-      </router-link> -->
+      <div class="navbar-item noselect">
+        <!-- svg is required here!! -->
+        <router-link class="navbar-item" to="/">
+          <img :src="require('@/assets/Logo_nobg.png')" alt="Localzon(e)" id="logo">
+        </router-link>
+
+      </div>
       <a role="button" class="navbar-burger" id="navbarBurger" aria-label="menu" aria-expanded="false"
          data-target="navbarLinks" @click="toggleMenuOnMobile">
         <span aria-hidden="true"></span>
@@ -15,24 +18,22 @@
 
     <div id="navbarLinks" class="navbar-menu">
       <div class="navbar-start">
-        <router-link v-for="route in routes" :key="route" :to="route.path" class="navbar-item noselect">
+        <router-link v-for="route in routes" :key="route" :to="route.path" class="navbar-item noselect localzone">
           {{ $t('views.' + route.name + '.name') }}
         </router-link>
       </div>
 
-      <div class="navbar-center">
-        <div class="navbar-item title noselect">
-          Localzon(e)
-        </div>
-      </div>
+      <!-- <div class="navbar-item">
+        <ticker />
+      </div> -->
 
 
       <div class="navbar-end">
-        <div class="navbar-item ">
+        <div class="navbar-item">
           <div class="select">
             <select v-model="$i18n.locale">
               <option v-for="locale in $i18n.availableLocales" :key="`locale-${locale}`" :value="locale">{{
-                  locale
+                  mapLocaleToLanguage[locale]
                 }}
               </option>
             </select>
@@ -52,9 +53,21 @@
 <script>
 export default {
   name: 'navigation-bar',
+  data () {
+    return {
+      mapLocaleToLanguage: {
+        'de': 'Deutsch',
+        'en': 'English'
+      },
+      notIncludedRoutes: [
+          'PageNotFound',
+          'SignIn'
+      ]
+    }
+  },
   computed: {
     routes: function () {
-      return this.$router.options.routes.filter(r => r.name !== 'PageNotFound').filter((v, i, a) => a.indexOf(v) === i)
+      return this.$router.options.routes.filter(r => !this.notIncludedRoutes.includes(r.name)).filter((v, i, a) => a.indexOf(v) === i)
     }
   },
   methods: {
@@ -76,35 +89,3 @@ export default {
 }
 
 </script>
-
-<style>
-.navbar {
-  background-color: var(--navbar-bg);
-}
-
-/* mobile */
-@media screen and (max-width: 1023px) {
-  .navbar {
-    color: black;
-  }
-}
-
-/* desktop */
-@media screen and (min-width: 1024px) {
-  .navbar-item, .navbar-link {
-    color: var(--navbar-color)
-  }
-
-  .navbar-center {
-    position: fixed;
-    left: 50%;
-    transform: translate(-50%, 0);
-    margin: auto;
-  }
-}
-
-#navbarBurger {
-  color: var(--navbar-color);
-}
-
-</style>