+++ src/components/signIn/signIn.vue
@@ -1,0 +1,5 @@
+<template>
+  <router-link class="button is-primary" to="/SignIn">
+    <strong>{{ $t('components.signIn.signIn_button') }}</strong>
+  </router-link>
+</template>