+++ src/functions/checkPassword.js
@@ -1,0 +1,5 @@
+export function checkPassword (password) {
+    // eslint-disable-next-line no-control-regex
+    const passwordRegex = /^(?=.*[a-z])(?=.*[A-Z])(?=.*\d)(?=.*[@$!%*?&])[A-Za-z\d@$!%*?&]{8,}$/
+    return passwordRegex.test(password)
+}