+++ vue.config.js
@@ -2,11 +2,14 @@ module.exports = {
     publicPath: process.env.NODE_ENV === 'production'
         ? '/website-frontend/'
         : '/',
-    chainWebpack: config => {
-        config.module
-            .rule('images')
-            .use('url-loader')
-            .loader('url-loader')
-            .tap(options => Object.assign(options, {limit: 10240}))
+
+    configureWebpack: {
+        optimization: {
+            splitChunks: {
+                minSize: 500,
+                maxSize: 250000
+            },
+            flagIncludedChunks: true
+        }
     }
 }